Move deposit and exchange apis def to the api dir
