+++ src/utils/index.ts
@@ -3,3 +3,4 @@ export * from './format'
 export * from './ethereum'
 export * from './miscellaneous'
 export * from './autoconnect'
+export * from './exchange'
